Migrate from Integer to BigInteger for UserIDs
Query to migrate the database itself:
ALTER TABLE afk_users ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE userinfo ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE userbio ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE antiflood ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE gbans ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE warns ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE connection ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE connection_history ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE human_checks ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE approval ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE users ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE chat_members ALTER COLUMN priv_chat_id SET DATA TYPE BigInt;
ALTER TABLE chat_members ALTER COLUMN "user" SET DATA TYPE BigInt;
ALTER TABLE royals ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE user_report_settings ALTER COLUMN user_id SET DATA TYPE BigInt;

+++ SkyzuRobot/modules/sql/welcome_sql.py
@@ -4,7 +4,8 @@ from typing import Union
 
 from SkyzuRobot.modules.helper_funcs.msg_types import Types
 from SkyzuRobot.modules.sql import BASE, SESSION
-from sqlalchemy import BigInteger, Boolean, Column, Integer, String, UnicodeText
+from sqlalchemy import Boolean, Column, Integer, String, UnicodeText
+from sqlalchemy.sql.sqltypes import BigInteger
 
 DEFAULT_WELCOME_MESSAGES = [
     "{first} is here!",  # Discord welcome messages copied
@@ -294,7 +295,7 @@ class WelcomeMute(BASE):
 
 class WelcomeMuteUsers(BASE):
     __tablename__ = "human_checks"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14), primary_key=True)
     human_check = Column(Boolean)
